requirements.txt
absl-py==0.9.0
amqp==1.4.9
anyjson==0.3.3
appnope==0.1.0
asgiref==3.2.7
astor==0.8.1
attrs==19.3.0
audioread==2.1.5
backcall==0.1.0
beautifulsoup4==4.6.0
billiard==3.3.0.23
bleach==1.5.0
bs4==0.0.1
cachetools==2.0.1
celery==3.1.24
certifi==2017.7.27.1
cffi==1.14.0
chardet==3.0.4
click==6.7
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9
cycler==0.10.0
decorator==4.1.2
defusedxml==0.6.0
dill==0.2.7.1
Django==3.0.5
django-rest-knox==4.1.0
djangorestframework==3.11.0
djangorestframework-files==1.1.0
drf-yasg==1.17.1
entrypoints==0.3
ffprobe==0.5
Flask==0.12.2
Flask-Cors==3.0.3
flex==6.14.1
future==0.16.0
gapic-google-cloud-datastore-v1==0.15.3
gapic-google-cloud-error-reporting-v1beta1==0.15.3
gapic-google-cloud-logging-v2==0.91.3
gapic-google-cloud-pubsub-v1==0.15.4
gapic-google-cloud-spanner-admin-database-v1==0.15.3
gapic-google-cloud-spanner-admin-instance-v1==0.15.3
gapic-google-cloud-spanner-v1==0.15.3
gast==0.2.2
google-api-core==1.1.2
google-auth==1.1.1
google-cloud==0.27.0
google-cloud-bigquery==0.26.0
google-cloud-bigtable==0.26.0
google-cloud-core==0.26.0
google-cloud-datastore==1.2.0
google-cloud-dns==0.26.0
google-cloud-error-reporting==0.26.0
google-cloud-language==0.27.0
google-cloud-logging==1.2.0
google-cloud-monitoring==0.26.0
google-cloud-pubsub==0.27.0
google-cloud-resource-manager==0.26.0
google-cloud-runtimeconfig==0.26.0
google-cloud-spanner==0.26.0
google-cloud-speech==0.28.0
google-cloud-storage==1.3.2
google-cloud-translate==1.1.0
google-cloud-videointelligence==0.25.0
google-cloud-vision==0.26.0
google-gax==0.15.16
google-pasta==0.2.0
google-resumable-media==0.3.0
googleapis-common-protos==1.5.3
grpc-google-iam-v1==0.11.4
grpcio==1.28.1
h5py==2.10.0
html5lib==0.9999999
httplib2==0.10.3
idna==2.6
importlib-metadata==1.6.0
inflect==0.2.5
inflection==0.4.0
ipdb==0.10.3
ipykernel==5.2.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
iso8601==0.1.12
itsdangerous==0.24
itypes==1.1.0
jamo==0.4.1
jedi==0.11.0
Jinja2==2.9.6
joblib==0.15.1
jsonpointer==2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
kombu==3.0.37
librosa==0.6.2
llvmlite==0.32.0
m3u8==0.3.3
Markdown==2.6.9
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
monotonic==1.3
mysqlclient==1.4.6
nbconvert==5.6.1
nbformat==5.0.5
nltk==3.2.5
notebook==6.0.3
numba==0.49.0
numpy==1.18.3
oauth2client==3.0.0
opt-einsum==3.2.1
packaging==20.3
pandocfilters==1.4.2
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
ply==3.8
prometheus-client==0.7.1
prompt-toolkit==1.0.15
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-error-reporting-v1beta1==0.15.3
proto-google-cloud-logging-v2==0.91.3
proto-google-cloud-pubsub-v1==0.15.4
proto-google-cloud-spanner-admin-database-v1==0.15.3
proto-google-cloud-spanner-admin-instance-v1==0.15.3
proto-google-cloud-spanner-v1==0.15.3
protobuf==3.12.2
ptyprocess==0.5.2
pyasn1==0.3.7
pyasn1-modules==0.1.5
pycparser==2.20
pydub==0.20.0
Pygments==2.2.0
pyparsing==2.2.0
pyreadline==2.1
pyrsistent==0.16.0
python-dateutil==2.6.1
pytz==2017.2
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.0
qtconsole==4.7.2
QtPy==1.9.0
requests==2.18.4
resampy==0.2.0
rfc3987==1.3.8
rsa==3.4.2
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scikit-learn==0.22.2.post1
scipy==1.4.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.3.1
strict-rfc3339==0.7
tenacity==4.4.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tinytag==0.18.0
tornado==6.0.4
tqdm==4.19.2
traitlets==4.3.2
Unidecode==1.0.22
uritemplate==3.0.1
urllib3==1.22
validate-email==1.3
vine==1.3.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
widgetsnbextension==3.5.1
wrapt==1.12.1
youtube-dl==2017.10.15.1
zipp==3.1.0
